refactor: reorganize project structure and move files to feature folders
- Move supabase config to dedicated config folder
- Organize components into feature-based folders (auth, budget, education, grocery, layout, user)
- Consolidate SQL schema files into supabase folder
- Update imports to reflect new file locations
