+++ components/education/VideoPlayer.tsx
