+++ components/user/UserProfile.tsx
@@ -1,7 +1,8 @@
 'use client'
 
 import { useState } from 'react'
-import { supabase } from '@/lib/supabase'
+import { supabase } from '@/lib/config/supabase'
+
 import { User } from '@supabase/supabase-js'
 
 interface UserProfileProps {
