+++ hooks/useAuth.ts
@@ -2,7 +2,8 @@
 
 import { useState, useEffect } from 'react'
 import { User } from '@supabase/supabase-js'
-import { supabase } from '@/lib/supabase'
+import { supabase } from '@/lib/config/supabase'
+
 
 export function useAuth() {
   const [user, setUser] = useState<User | null>(null)
