+++ supabase/budget-schema.sql
