+++ supabase/supabase-education-schema.sql
